vscode

+++ setup.sh
@@ -77,7 +77,9 @@ setup.vscode() {
   local vscode_user_dir="${HOME}/Library/Application Support/Code/User"
   if [[ -d "$vscode_user_dir" ]]; then
     for file in settings.json keybindings.json; do
-      mv "${vscode_user_dir}/${file}" "${vscode_user_dir}/${file}.bk"
+      if [[ -f "${vscode_user_dir}/${file}" ]]; then
+        mv "${vscode_user_dir}/${file}" "${vscode_user_dir}/${file}.bk"
+      fi
       ln -s "${HOME}/dotfiles/vscode/${file}" "${vscode_user_dir}/${file}"
     done
   fi
